Fix the braindead isspace() insanity
<ctype.h> functions are insane; they cannot be simply invoked as
  char c = read_byte();
  int is_it_a_space = isspace(c);
Because they take an 'int' parameter, not 'char', and expect that parameter
to be either non-negative (and in the range of 'unsigned char') or EOF,
which is some implementation-defined negative value. Passing any other
value, according to the standard, leads to undefined behavior.

The signedness of 'char' ('unsigned char', 'signed char' and 'char' are
three different types!) depends on the implementation. And if it's signed,
you're fucked.

glibc "support[s] negative `signed char' values for broken old programs",
which is good until you consider single-byte character sets and note that
it confuses EOF (-1 in glibc) with a legitimate character 0xFF (e.g. in
CP1251 it's "я", a letter of the Russian alphabet).